// common/src/main/ets/model/response/PlayUrlResponse.ts
/**
 * Copyright (c) 2023 Wathinst <wxz@xkzhineng.com>
 * OhBili is licensed under Mulan PSL v2.
 * You can use this software according to the terms and conditions of the Mulan PSL v2.
 * You may obtain a copy of Mulan PSL v2 at:
 * http://license.coscl.org.cn/MulanPSL2
 * THIS SOFTWARE IS PROVIDED ON AN "AS IS" BASIS, WITHOUT WARRANTIES OF ANY KIND,
 * EITHER EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO NON-INFRINGEMENT,
 * MERCHANTABILITY OR FIT FOR A PARTICULAR PURPOSE.
 * See the Mulan PSL v2 for more details.
 */

import { BaseResponse } from './Response';

export interface PlayUrlResponse extends BaseResponse<PlayUrl> {

}

export interface PlayUrl {
    quality: number
    format: string
    seek_param: string
    seek_type: string
    durl: PlayUrlDurl[]
}

export interface PlayUrlDurl {
    length: number
    size: number
    url: string
}